// src/types/stem.d.ts
export interface IStemStruct {
  stem_down_y_base_offset: number;
  stem_up_y_base_offset: number;
  stem_down_y_offset: number;
  stem_up_y_offset: number;
  stemletHeight: number;
  isStemlet: boolean;
  hide: boolean;
  stem_direction: number;
  stem_extension: number;
  y_bottom: number;
  y_top: number;
  x_end: number;
  x_begin: number;
}